+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/custom.py
@@ -1186,8 +1186,8 @@ def list_extensions(cmd, resource_group_name, vm_name):
     return result
 
 
-def set_extension(cmd, resource_group_name, vm_name, vm_extension_name, publisher,
-                  version=None, settings=None, protected_settings=None, no_auto_upgrade=False, force_update=False):
+def set_extension(cmd, resource_group_name, vm_name, vm_extension_name, publisher, version=None, settings=None,
+                  protected_settings=None, no_auto_upgrade=False, force_update=False, no_wait=False):
     vm = get_vm(cmd, resource_group_name, vm_name, 'instanceView')
     client = _compute_client_factory(cmd.cli_ctx)
 
@@ -1203,7 +1203,8 @@ def set_extension(cmd, resource_group_name, vm_name, vm_extension_name, publishe
                                   auto_upgrade_minor_version=(not no_auto_upgrade))
     if force_update:
         ext.force_update_tag = str(_gen_guid())
-    return client.virtual_machine_extensions.create_or_update(resource_group_name, vm_name, instance_name, ext)
+    return sdk_no_wait(no_wait, client.virtual_machine_extensions.create_or_update,
+                       resource_group_name, vm_name, instance_name, ext)
 # endregion
 
 
@@ -2361,7 +2362,8 @@ def list_vmss_extensions(cmd, resource_group_name, vmss_name):
 
 
 def set_vmss_extension(cmd, resource_group_name, vmss_name, extension_name, publisher, version=None,
-                       settings=None, protected_settings=None, no_auto_upgrade=False, force_update=False):
+                       settings=None, protected_settings=None, no_auto_upgrade=False, force_update=False,
+                       no_wait=False):
     client = _compute_client_factory(cmd.cli_ctx)
     vmss = client.virtual_machine_scale_sets.get(resource_group_name, vmss_name)
     VirtualMachineScaleSetExtension, VirtualMachineScaleSetExtensionProfile = cmd.get_models(
@@ -2390,7 +2392,8 @@ def set_vmss_extension(cmd, resource_group_name, vmss_name, extension_name, publ
         vmss.virtual_machine_profile.extension_profile = VirtualMachineScaleSetExtensionProfile(extensions=[])
     vmss.virtual_machine_profile.extension_profile.extensions.append(ext)
 
-    return client.virtual_machine_scale_sets.create_or_update(resource_group_name, vmss_name, vmss)
+    return sdk_no_wait(no_wait, client.virtual_machine_scale_sets.create_or_update,
+                       resource_group_name, vmss_name, vmss)
 # endregion